Fix error where the exception was hidden in the event autopep8 failed
If there was an internal error running autopep8, it would raise an
exception and close the stdin.

Sending the EOF would then fail with an unhelpful message about
the file being closed, worse - the actual exception message would not
be shown making it difficult to debug the cause of the error.

Now file errors are accounted for.